added planes

+++ include/Object.h
@@ -5,7 +5,7 @@
 
 struct HitRecord {
     bool hit = false;
-    float t = 1e9f;
+    double t = 1e9f;
     Vec3 point;
     Vec3 normal;
     Material material;
@@ -19,5 +19,5 @@ public:
     
     Material material;
     // Abstract hit method for intersection logic
-    virtual bool hit(const Ray& ray, float t_min, HitRecord& rec) const = 0;
+    virtual bool hit(const Ray& ray, double t_min, HitRecord& rec) const = 0;
 };
